Stop staging tasks during shutdown sequence.
This change ensures that in-flight staging tasks are stopped during the DEA's
shutdown sequence. The cloud controllers (ng) are sent error responses
via NATS for all in-flight staging tasks being stopped. This ensures that
the stopped tasks will eventually be retried (possibly on different DEAs).

This commit also includes an rspec bump.